# fe/vdbas_exp_fe/apiContract/API-contract.yaml
openapi: 3.0.3
info:
  title: EXP CAPEX Dossier API
  description: |
    API quản lý **Hồ sơ Chi đầu tư (CAPEX)** – Phân hệ Chi (EXP) trong hệ thống VDBAS.

    ## Luồng trạng thái hồ sơ
    ```
    Start → DRAFT → SAVED → SUBMITTED → APPROVED (Checker) → APPROVED (Approver) → COMPLETED
                                ↘ REJECTED (Checker reject / Approver reject)
               ↘ CANCELLED (Maker xoá)
    ```

    ## Phân quyền
    - **Maker**: Tạo mới, Sửa, Xoá, Lưu nháp, Gửi kiểm soát (trong phạm vi TREASURY_CODE)
    - **Checker**: Kiểm soát (Approve/Reject) hồ sơ ở trạng thái SUBMITTED
    - **Approver**: Phê duyệt (Approve/Reject) hồ sơ ở trạng thái APPROVED (đã kiểm soát)
    - **Viewer**: Chỉ xem

    ## Lưu ý kỹ thuật
    - Mọi ID kiểu UUID (chuỗi 36 ký tự, dạng `xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx`)
    - Ngày dùng `yyyy-MM-dd`; DateTime dùng ISO 8601 `yyyy-MM-ddTHH:mm:ss+07:00`
    - Số tiền VND kiểu `integer` (đơn vị VND)
    - Optimistic lock qua trường `version` (gửi kèm khi Update)
    - Phòng double-submit: header `X-Idempotency-Key` (UUID do client sinh)
    - Phân quyền theo `TREASURY_CODE` lấy từ JWT claim
  version: "0.2.0"
  contact:
    name: VDBAS Development Team
servers:
  - url: /api/v1
    description: API version 1
security:
  - bearerAuth: []
...
paths:
...
  /exp/capex/dossiers/{dossierId}:
    parameters:
      - $ref: '#/components/parameters/DossierId'

    get:
      tags: [Dossiers]
      operationId: getDossier
      summary: Xem chi tiết hồ sơ (EXP.CAPEX_DOSSIER.VIEW.OPEN)
      description: |
        Trả về đầy đủ thông tin hồ sơ: thông tin chung (§B1.1), danh sách chứng từ (§B1.2),
        đính kèm, lịch sử giao dịch, trạng thái phê duyệt.
        Tất cả vai trò có quyền theo TREASURY_CODE.
      responses:
        '200':
          description: Chi tiết hồ sơ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Dossiers]
      operationId: updateDossier
      summary: Cập nhật hồ sơ (EXP.CAPEX_DOSSIER.EDIT.SAVE)
      description: |
        Chỉ áp dụng khi F_STATUS ∈ {DRAFT, SAVED} và NSD là Maker gốc.
        Optimistic lock: `version` phải khớp với DB (VAL-15); F-VER tăng +1 sau khi lưu.
        Ghi audit oldValue→newValue. Trạng thái không đổi.
        Yêu cầu vai trò Maker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DossierUpdateRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DossierUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Optimistic lock conflict (VAL-15 / VDBAS-EXP-0005)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errorCode: VDBAS-EXP-0005
                message: "Bản ghi đã bị thay đổi từ phiên khác. Vui lòng tải lại trước khi tiếp tục"
                traceId: "abc-123"

    delete:
      tags: [Dossiers]
      operationId: deleteDossier
      summary: Xoá mềm hồ sơ → CANCELLED (EXP.CAPEX_DOSSIER.DELETE.CONFIRM)
      description: |
        Soft-delete: STATUS=0, F_STATUS=CANCELLED. Ghi audit. Trả về MSG-OK-DELETE.
        Chỉ áp dụng khi F_STATUS ∈ {DRAFT, SAVED} và NSD là Maker gốc.
        Yêu cầu vai trò Maker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteDossierRequest'
      responses:
        '200':
          description: Xoá thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                errorCode: VDBAS-EXP-SUC03
                message: "Xoá hồ sơ thành công"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Trạng thái không cho phép xoá (VAL-13 / VDBAS-EXP-0003)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════
  #  WORKFLOW – GỬI KIỂM SOÁT
  # ═══════════════════════════════════════════
...
  /exp/capex/dossiers/{dossierId}/documents:
    parameters:
      - $ref: '#/components/parameters/DossierId'

    get:
      tags: [Documents]
      operationId: listDossierDocuments
      summary: Danh sách chứng từ trong hồ sơ (§B1.2)
      description: |
        Trả về danh sách chứng từ gắn với hồ sơ. Bao gồm tổng tiền footer.
        Hiển thị sau khi hồ sơ đã được tạo (không hiện ở NEW form).
      responses:
        '200':
          description: Danh sách chứng từ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Documents]
      operationId: addDocumentToDossier
      summary: Thêm chứng từ vào hồ sơ (EXP.CAPEX_DOSSIER.NEW.ADD_DOC)
      description: |
        Liên kết chứng từ mới vào hồ sơ. Chứng từ phải được tạo theo chức năng
        EXP.CAPEX_DOSSIER.3.3.1 (Giấy ĐNTT) hoặc loại tương ứng.
        Hồ sơ phải ở trạng thái DRAFT hoặc SAVED. Yêu cầu vai trò Maker.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDocumentRequest'
      responses:
        '201':
          description: Thêm chứng từ thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /exp/capex/dossiers/{dossierId}/documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DossierId'
      - name: documentId
        in: path
        required: true
        description: UUID chứng từ
        schema: { type: string, format: uuid }

    get:
      tags: [Documents]
      operationId: getDossierDocument
      summary: Xem chi tiết chứng từ
      responses:
        '200':
          description: Chi tiết chứng từ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Documents]
      operationId: updateDossierDocument
      summary: Cập nhật chứng từ trong hồ sơ
      description: |
        Chỉ cập nhật khi hồ sơ ở trạng thái DRAFT/SAVED và NSD là Maker gốc.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Documents]
      operationId: removeDossierDocument
      summary: Xoá chứng từ khỏi hồ sơ
      description: |
        Tách liên kết chứng từ khỏi hồ sơ (soft-delete).
        Chỉ khi hồ sơ ở DRAFT/SAVED và NSD là Maker gốc.
      responses:
        '200':
          description: Xoá thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ═══════════════════════════════════════════
  #  ATTACHMENTS – ĐÍNH KÈM HỒ SƠ
  # ═══════════════════════════════════════════
  /exp/capex/dossiers/{dossierId}/attachments:
    parameters:
      - $ref: '#/components/parameters/DossierId'

    get:
      tags: [Attachments]
      operationId: listDossierAttachments
      summary: Danh sách file đính kèm hồ sơ (EXP.CAPEX_DOSSIER.VIEW.ATTACHMENTS)
      responses:
        '200':
          description: Danh sách đính kèm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Attachments]
      operationId: uploadDossierAttachment
      summary: Upload file đính kèm hồ sơ (EXP.CAPEX_DOSSIER.ATTACH.UPLOAD)
      description: |
        Upload file đính kèm cho hồ sơ. Yêu cầu:
        - ≤ 10MB/file (VAL-09 / VDBAS-VAL-0014)
        - Định dạng: pdf, docx, xlsx, jpg, png (P1-4)
        - Kiểm tra MIME type + magic bytes
        - Quét virus trước khi lưu
        - Cảnh báo trùng lặp (cùng tên + dung lượng): VAL-20
        Yêu cầu vai trò Maker; hồ sơ ở DRAFT/SAVED.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, attachmentTypeCode]
              properties:
                file:
                  type: string
                  format: binary
                  description: File đính kèm (≤ 10MB; pdf/docx/xlsx/jpg/png)
                attachmentTypeCode:
                  type: string
                  description: "Loại đính kèm: CHUNG_TU_GOC | HOP_DONG | HOA_DON | BANG_KE | VAN_BAN_KHAC"
                  enum: [CHUNG_TU_GOC, HOP_DONG, HOA_DON, BANG_KE, VAN_BAN_KHAC]
                description:
                  type: string
                  maxLength: 250
                  description: Mô tả file (không bắt buộc)
      responses:
        '201':
          description: Upload thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: File vượt quá 10MB (VDBAS-VAL-0014)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Định dạng file không hỗ trợ (VDBAS-VAL-0015)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /exp/capex/dossiers/{dossierId}/attachments/{attachmentId}:
    parameters:
      - $ref: '#/components/parameters/DossierId'
      - name: attachmentId
        in: path
        required: true
        description: UUID file đính kèm
        schema: { type: string, format: uuid }

    get:
      tags: [Attachments]
      operationId: downloadDossierAttachment
      summary: Tải xuống file đính kèm (EXP.CAPEX_DOSSIER.ATTACH.DOWNLOAD)
      description: Ghi audit truy cập khi download (BIZ-007).
      responses:
        '200':
          description: Nội dung file
          headers:
            Content-Disposition:
              schema: { type: string }
              example: attachment; filename="document.pdf"
            Content-Type:
              schema: { type: string }
              example: application/pdf
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Attachments]
      operationId: deleteDossierAttachment
      summary: Xoá file đính kèm (EXP.CAPEX_DOSSIER.ATTACH.DELETE)
      description: Chỉ khi hồ sơ ở DRAFT/SAVED và NSD là Maker gốc.
      responses:
        '200':
          description: Xoá thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ═══════════════════════════════════════════
  #  AUDIT – LỊCH SỬ PHÊ DUYỆT
  # ═══════════════════════════════════════════
...
  /exp/capex/dossiers/{dossierId}/audit-log:
    parameters:
      - $ref: '#/components/parameters/DossierId'
    get:
      tags: [Audit]
      operationId: getDossierAuditLog
      summary: Lịch sử thay đổi / Audit (EXP.CAPEX_DOSSIER.VIEW.HISTORY)
      description: |
        Trả về lịch sử thao tác: CREATED_BY, CREATED_DATE, UPDATED_BY, UPDATED_DATE,
        action, oldValue→newValue, IP. Sort DESC theo thời gian (mới nhất lên đầu).
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lịch sử audit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════
  #  EXPORT – KẾT XUẤT
  # ═══════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT chứa claims userId, role, treasuryCode
  parameters:
    DossierId:
      name: dossierId
      in: path
      required: true
      description: UUID hồ sơ (EXP_DOSSIER.ID)
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      description: UUID do client sinh – phòng chống double-submit
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            errorCode: VDBAS-VAL-0002
            message: "Dữ liệu đầu vào không hợp lệ"
            traceId: "xyz-456"
            fieldErrors:
              - field: projectCode
                errorCode: VDBAS-VAL-0002
                message: "Vui lòng nhập Mã dự án/công trình"
    Unauthorized:
      description: Chưa xác thực (VDBAS-AUT-0001)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errorCode: VDBAS-AUT-0002
            message: "Phiên đăng nhập đã hết hạn, vui lòng đăng nhập lại"
            traceId: "xyz-789"
    Forbidden:
      description: Không có quyền thực hiện (VDBAS-AUT-0001)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errorCode: VDBAS-AUT-0001
            message: "Bạn không có quyền thực hiện thao tác này"
            traceId: "abc-789"
    NotFound:
      description: Không tìm thấy bản ghi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Vi phạm quy tắc nghiệp vụ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DossierStatus:
      type: string
      description: |
        Trạng thái hồ sơ (enum chuẩn §A11.1):
        - DRAFT: Lưu nháp (badge xám)
        - SAVED: Đã lưu (badge xám)
        - VALIDATED: Đã kiểm tra — nhận tự động (badge xanh dương)
        - SUBMITTED: Đã gửi phê duyệt (badge xanh dương)
        - APPROVED: Đã phê duyệt / Đã kiểm soát (badge xanh lá)
        - REJECTED: Đã từ chối (badge đỏ)
        - COMPLETED: Đã hoàn thành (badge xanh lá)
        - CANCELLED: Đã huỷ (badge cam)
      enum:
        - DRAFT
        - SAVED
        - VALIDATED
        - SUBMITTED
        - APPROVED
        - REJECTED
        - COMPLETED
        - CANCELLED
...
    DossierUpdateRequest:
      type: object
      required: [version]
      description: |
        Cập nhật hồ sơ. Trường `version` bắt buộc để kiểm tra optimistic lock.
        Các trường immutable (DOSSIER_CODE, CREATED_BY, CREATED_DATE) không được gửi — backend sẽ reject.
      properties:
        version:
          type: integer
          description: "Phiên bản F-VER hiện tại (optimistic lock — VAL-15)"
          example: 1
        sendDate:
          type: string
          format: date
        projectCode:
          type: string
          maxLength: 100
        projectSpecificCode:
          type: string
          maxLength: 100
          nullable: true
        investorCode:
          type: string
          maxLength: 100
        projectManagementCode:
          type: string
          maxLength: 100
    DeleteDossierRequest:
      type: object
      required:
        - deleteReason
        - confirmReviewed
      properties:
        deleteReason:
          type: string
          minLength: 10
          maxLength: 500
          description: "Lý do xoá — bắt buộc ≥ 10 ký tự (VAL-16)"
          example: "Hồ sơ nhập sai thông tin dự án, cần tạo lại từ đầu"
        confirmReviewed:
          type: boolean
          description: "Xác nhận đã rà soát trước khi xoá — bắt buộc = true (VAL-16)"
          example: true
...
    AddDocumentRequest:
      type: object
      required:
        - documentTypeCode
        - documentNo
        - documentDate
        - accountingDate
        - baseAmount
      properties:
        documentTypeCode:
          type: string
          maxLength: 100
          description: "Mã loại chứng từ (LOV.Loại chứng từ)"
          example: "GIAY_DNTT"
        documentName:
          type: string
          maxLength: 500
          description: Tên chứng từ (auto-fill theo loại, có thể override)
        documentNo:
          type: string
          maxLength: 200
          description: "Số chứng từ (format: EXP/CAPEX/YYYY/XXXXX)"
          example: "EXP/CAPEX/2025/00001"
        documentDate:
          type: string
          format: date
          description: Ngày chứng từ
        accountingDate:
          type: string
          format: date
          description: Ngày hạch toán (phải nằm trong kỳ kế toán mở — VAL-08)
        originalAmount:
          type: integer
          nullable: true
          description: "Số tiền nguyên tệ (để null nếu VND)"
        baseAmount:
          type: integer
          description: Số tiền VND
          example: 500000000
    UpdateDocumentRequest:
      allOf:
        - $ref: '#/components/schemas/AddDocumentRequest'
        - type: object
          required: [version]
          properties:
            version:
              type: integer
              description: Optimistic lock version
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        errorCode:
          type: string
          description: Mã thông báo thành công (VDBAS-EXP-SUCxx)
          example: VDBAS-EXP-SUC01
        message:
          type: string
          example: "Lưu hồ sơ thành công"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errorCode:
          type: string
          description: "Mã lỗi chuẩn (VDBAS-VAL-*, VDBAS-EXP-*, VDBAS-SYS-*, VDBAS-AUT-*)"
          example: VDBAS-EXP-0003
        message:
          type: string
          example: "Hồ sơ đang ở trạng thái SUBMITTED, không cho phép Sửa/Xoá"
        traceId:
          type: string
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Tên trường lỗi
                    example: projectCode
                  errorCode:
                    type: string
                    example: VDBAS-VAL-0002
                  message:
                    type: string
                    example: "Vui lòng nhập Mã dự án/công trình"
...
    DossierDetail:
      type: object
      description: Thông tin đầy đủ hồ sơ (§B1.1 + §B1.2 + metadata)
      properties:
        id:
          type: string
          format: uuid
        dossierCode:
          type: string
          description: "Mã hồ sơ (auto-gen, immutable — VAL-17)"
          example: "EXP/CAPEX/2025/00001"
        version:
          type: integer
          description: "Phiên bản F-VER (optimistic lock)"
          example: 2
        # Thông tin kho bạc (từ JWT scope)
        treasuryCode:
          type: string
          example: "0012"
        treasuryName:
          type: string
          example: "Kho bạc nhà nước khu vực I - PGD số 12"
        # Thông tin giao dịch
        sendDate:
          type: string
          format: date
        dataSourceCode:
          type: string
          enum: [THU_CONG, DVC]
        dataSourceName:
          type: string
          example: "Thủ công"
        fStatus:
          $ref: '#/components/schemas/DossierStatus'
        fStatusName:
          type: string
          example: "Đã lưu"
        # Thông tin dự án
        projectCode:
          type: string
          example: "7122155"
        projectName:
          type: string
          example: "Dự án nâng cấp bệnh viện Bạch Mai"
        projectType:
          type: string
          enum: [MILITARY, CITIZEN]
        projectSpecificCode:
          type: string
          nullable: true
        projectSpecificName:
          type: string
          nullable: true
        investorCode:
          type: string
          example: "1059227"
        investorName:
          type: string
          example: "Thông tấn xã Việt Nam"
        projectManagementCode:
          type: string
          example: "3029123"
        projectManagementName:
          type: string
          example: "BQLDA bệnh viện Bạch Mai"
        # Backend-managed (§B5 — Out of scope for UI input)
        assignUser:
          type: string
          description: User đang giữ xử lý (backend cập nhật tự động)
        sla:
          type: string
          format: date-time
          description: Thời hạn xử lý SLA
        completedDate:
          type: string
          format: date
          nullable: true
        # Audit fields
        createdBy:
          type: string
        createdDate:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedDate:
          type: string
          format: date-time
        # Danh sách chứng từ (§B1.2 grid)
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        totalBaseAmount:
          type: integer
          description: "Tổng tiền VND (footer grid §B1.2)"
    DocumentSummary:
      type: object
      description: Thông tin tóm tắt chứng từ — hiển thị trên grid §B1.2
      properties:
        id:
          type: string
          format: uuid
        seqNo:
          type: integer
          description: STT (auto, bắt đầu từ 1)
        documentTypeCode:
          type: string
        documentName:
          type: string
          example: "Giấy đề nghị thanh toán vốn mẫu số 04.a.nn/TT"
        documentNo:
          type: string
          example: "EXP/CAPEX/2025/00001"
        documentDate:
          type: string
          format: date
        accountingDate:
          type: string
          format: date
        originalAmount:
          type: integer
          nullable: true
          description: "Số tiền nguyên tệ (null = VND)"
        baseAmount:
          type: integer
          description: Số tiền VND
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          properties:
            dossierId:
              type: string
              format: uuid
            status:
              type: integer
              enum: [0, 1]
              description: "Hiệu lực: 0=Không hiệu lực, 1=Hiệu lực"
            version:
              type: integer
            createdBy:
              type: string
            createdDate:
              type: string
              format: date-time
            updatedBy:
              type: string
            updatedDate:
              type: string
              format: date-time
    AttachmentInfo:
      type: object
      description: Thông tin file đính kèm (§B3.3)
      properties:
        id:
          type: string
          format: uuid
        attachmentTypeCode:
          type: string
          enum: [CHUNG_TU_GOC, HOP_DONG, HOA_DON, BANG_KE, VAN_BAN_KHAC]
        attachmentTypeName:
          type: string
          example: "Chứng từ gốc"
        fileName:
          type: string
          example: "hop-dong-2025-001.pdf"
        fileType:
          type: string
          example: "pdf"
        fileSize:
          type: integer
          description: Kích thước file (bytes)
          example: 2048576
        fileSizeDisplay:
          type: string
          description: Hiển thị KB/MB
          example: "2.0 MB"
        description:
          type: string
          nullable: true
        createdBy:
          type: string
        createdDate:
          type: string
          format: date-time
...
    AuditLogEntry:
      type: object
      description: Một dòng audit (EXP_AUDIT_LOG + §B3.2)
      properties:
        id:
          type: string
          format: uuid
        actionType:
          type: string
          enum: [INSERT, UPDATE, DELETE, APPROVE, REJECT]
        actionTimestamp:
          type: string
          format: date-time
        userId:
          type: string
        userDisplayName:
          type: string
        ipAddress:
          type: string
        tableName:
          type: string
        recordId:
          type: string
        oldValue:
          type: object
          nullable: true
          description: Giá trị cũ (JSON)
        newValue:
          type: object
          description: Giá trị mới (JSON)

    # ─── LIST/PAGED RESPONSES ─────────────────
...
    DossierUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        errorCode:
          type: string
          example: VDBAS-EXP-SUC01
        message:
          type: string
          example: "Lưu hồ sơ thành công"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            dossierCode:
              type: string
            fStatus:
              $ref: '#/components/schemas/DossierStatus'
            version:
              type: integer
              description: F-VER mới sau khi cập nhật (+1)
    DossierDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/DossierDetail'
...
    DocumentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/DocumentSummary'
            totalBaseAmount:
              type: integer
              description: "Tổng cộng (footer §B1.2)"
            totalOriginalAmount:
              type: integer
              nullable: true
    DocumentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/DocumentDetail'
    AttachmentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentInfo'
    AttachmentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/AttachmentInfo'
...
    AuditLogResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AuditLogEntry'
            pagination:
              $ref: '#/components/schemas/Pagination'

    # ─── LOV SCHEMAS ─────────────────────────
...